+++ sdk/rust/nym-sdk/src/error.rs
@@ -5,10 +5,13 @@ use std::path::PathBuf;
 pub enum Error {
     #[error("i/o error: {0}")]
     IoError(#[from] std::io::Error),
+
     #[error("toml serialization error: {0}")]
     TomlSerializationError(#[from] toml::ser::Error),
+
     #[error("toml deserialization error: {0}")]
     TomlDeserializationError(#[from] toml::de::Error),
+
     #[error("Ed25519 error: {0}")]
     Ed25519RecoveryError(#[from] nym_crypto::asymmetric::identity::Ed25519RecoveryError),
 
@@ -17,22 +20,30 @@ pub enum Error {
 
     #[error("key file encountered that we don't want to overwrite: {0}")]
     DontOverwrite(PathBuf),
+
     #[error("shared gateway key file encountered that we don't want to overwrite: {0}")]
     DontOverwriteGatewayKey(PathBuf),
+
     #[error("no gateway config available for writing")]
     GatewayNotAvailableForWriting,
+
     #[error("expected to received a directory, received: {0}")]
     ExpectedDirectory(PathBuf),
 
     #[error("failed to transition to registered state before connection to mixnet")]
     FailedToTransitionToRegisteredState,
+
     #[error(
         "registering with gateway when the client is already in a registered state is not \
          supported, and likely and user mistake"
     )]
     ReregisteringGatewayNotSupported,
+
     #[error("no gateway key set")]
     NoGatewayKeySet,
+
+    #[error("not socks5 configuration set")]
+    NoSocks5ConfigSet,
 
     #[error("failed to create reply storage backend: {source}")]
     StorageError {
